requirements.txt
absl-py==0.1.10
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
certifi==2018.4.16
cycler==0.10.0
Cython==0.28.2
decorator==4.3.0
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
Keras==2.1.6
kiwisolver==1.0.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.3
opencv-python==3.4.1.15
pandas==0.23.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
scikit-image==0.13.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Theano==1.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
snetd-alpha==0.1.0
jsonrpcserver==3.5.4
jsonrpcclient==2.5.2
aiohttp==3.2.1 
opencv-python==3.4.1.15
dlib==19.13.1
pytest
